Python/requirements.txt
ansible-core>=2.14.0
ansible-posix>=1.4.0
community.general>=5.8.0
amazon.aws>=4.0.0
azure.azcollection>=1.12.0
google.cloud>=0.34.0
boto3>=1.26.0
google-cloud-compute>=1.12.0
azure-identity>=1.12.0
pytest>=7.3.1
pytest-mock>=3.10.0
pyyaml>=6.0
python-dotenv>=1.0.0 